+++ docs/changelog.md
@@ -54,6 +54,10 @@
 
 ## 更新说明
 
+1.9.34 fix: fix stock_zh_a_tick_tx_js interface
+    
+    1. 修复 stock_zh_a_tick_tx_js 接口
+
 1.9.33 add: add stock_hk_main_board_spot_em interface
 
     1. 新增 stock_hk_main_board_spot_em 接口
@@ -2095,6 +2099,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.9.34 fix: fix stock_zh_a_tick_tx_js interface
 
 1.9.33 add: add stock_hk_main_board_spot_em interface